http://slashie.net/mantis/view.php?id=511

+++ src/net/slashie/expedition/domain/GoodsCache.java
@@ -165,7 +165,7 @@ public class GoodsCache extends AbstractFeature implements FoodConsumer{
 		return getItems();
 	}
 	
-	public int getFoodConsumptionMultiplier() {
+	public double getFoodConsumptionMultiplier() {
 		return 1;
 	}
